feat: study create, remove, share 구현 fix: develop branch 오류 수정

+++ backend/lighthouse/src/test/java/com/ssafy/lighthouse/domain/study/service/SessionServiceTest.java
@@ -15,7 +15,7 @@ public class SessionServiceTest {
 	@Autowired
 	private SessionService sessionService;
 
-	private static final int STUDY_ID = 1;
+	private static final Long STUDY_ID = 1L;
 
 	@Test
 	public void createSessionTest() {
